requirements.txt
SQLAlchemy
flask
python-dotenv
Flask-WTF
flask-login